change AuroraInfo states to objest, and fix recalling getLocations.
